requirements.txt
scikit-learn
altgraph==0.17.3
audioread==3.0.0
certifi==2022.12.07
cffi==1.15.1
cryptography==3.4.6
einops==0.6.0
future==0.18.3
julius==0.2.7
kthread==0.2.3
librosa
llvmlite
matchering==2.0.6
ml_collections==0.1.1
natsort==8.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
Pillow==9.3.0
psutil==5.9.4
pydub==0.25.1
pyglet==1.5.23
pyperclip==1.8.2
pyrubberband==0.3.0
pytorch_lightning==2.0.0
PyYAML==6.0
resampy==0.4.2
scipy==1.9.3
soundstretch==1.2
urllib3
wget==3.2
samplerate==0.1.0
screeninfo==0.8.1
diffq
playsound
onnx
onnxruntime
onnxruntime-gpu
onnx2pytorch
soundFile; sys_platform != 'darwin'
PySoundFile==0.9.0.post1; sys_platform == 'darwin'
numpy
###
--extra-index-url https://download.pytorch.org/whl/cu121 torch==2.3.1+cu121 torchvision==0.18.1+cu121 torchaudio==2.3.1+cu121
whisperx @ git+https://github.com/m-bain/whisperx.git
ass
srt
pylrc
natsort
python-dotenv
ffmpeg-python>=0.2.0
gradio==4.29.0
babel
colorama
multiprocess
yt_dlp
langdetect
## For fastapi and music separator webUI server
fastapi
uvicorn
starlette
itsdangerous
websockets
vamp
chord_extractor @ git+https://github.com/kennytat/chord-extractor.git
demucs==4.0.1